generalize player not found and already killed
