+++ messages.json
@@ -18,7 +18,7 @@
   "ONGOING_GAME_FOUND": "There's an ongoing game in this server, finish it first",
   "PLAYING_USER_FOUND": "{playing_users} currently playing the game on other channel",
   "ONGOING_GAME_NOT_FOUND": "There's no ongoing game in this server",
-  "ELIMINATED_PLAYER_NOT_FOUND": "Elimination failed because {player} is not playing in this channel",
-  "ELIMINATED_PLAYER_ALREADY_KILLED": "Elimination failed because {player} has been eliminated before",
+  "PLAYER_NOT_FOUND": "Command failed: {player} is not playing in this channel",
+  "PLAYER_ALREADY_KILLED": "Command failed: {player} has been eliminated before",
   "NOT_IN_GUESSING_TURN": "It's not yet the time for guessing"
 }
